requirements.txt
python-telegram-bot==20.7
git+https://github.com/rabilrbl/deepseek-api.git
